Add starry globe backdrop and polished modal
